issue data offset

+++ app/Http/Controllers/PengembanganController.php
@@ -6,39 +6,64 @@
 */
 namespace App\Http\Controllers;
 
+use App\Models\Dosen;
 use App\Models\Pengembangan;
 use App\Models\Periode;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+
+use function PHPUnit\Framework\isEmpty;
 
 class PengembanganController extends Controller
 {
     public function index()
     {
-        // $data = DB::table('pengembangans')
-        //     ->join('periodes', 'periodes.id', '=', 'pengembangans.periode_id')
-        //     ->join('jenis_pengdiris', 'jenis_pengdiris.id', '=', 'pengembangans.jenis_pengdiri_id')
-        //     ->get();
-        //     dd($data);
-        $data = Pengembangan::all();
+        $periode = DB::table('periodes')->first();
+        $user = Auth::user()->username;
+        if (Auth::user()->role_id !== 3) {
+            $data = Pengembangan::orderBy('created_at')->get();
+            dd($data);
+        } else {
+            $dosen = Dosen::where('user_id', $user)->first();
+            // $dosen = DB::table('dosens')
+            //     ->where('user_id', '=', $user)
+            //     ->get();
+             $data = Pengembangan::where('dosen_id', isset($dosen[0]->id))->get();
+
+        }
+        // $datas = Pengembangan::all();
         return view('pengembangan.index', compact('data'));
     }
 
     public function create()
     {
+        $this->authorize('create', Pengembangan::class);
+
         $periode        = DB::table('periodes')->get();
         $jenis_pengdiri = DB::table('jenis_pengdiris')->get();
-        return view('pengembangan.add', compact('periode', 'jenis_pengdiri'));
+        $dosen = User::join('dosens', 'users.username', '=', 'dosens.user_id')
+            ->where('dosens.status', 'aktif')
+            ->orderBy('dosens.created_at', 'desc')
+            ->get();
+        return view('pengembangan.add', compact('periode', 'jenis_pengdiri', 'dosen'));
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', Pengembangan::class);
+
         $this->validate($request, [
+            'dosen_id'          => 'required',
+            'periode_id'        => 'required',
+            'jenis_pengdiri_id' => 'required',
             'judul_pengdiri'    => 'required',
             'lokasi_pengdiri'   => 'required',
         ]);
 
         $data = Pengembangan::create([
+            'dosen_id'          => $request->dosen_id,
             'periode_id'        => $request->periode_id,
             'jenis_pengdiri_id' => $request->jenis_pengdiri_id,
             'judul_pengdiri'    => $request->judul_pengdiri,
@@ -52,6 +77,8 @@ class PengembanganController extends Controller
 
     public function edit(Pengembangan $pengembangan)
     {
+        $this->authorize('update', Pengembangan::class);
+
         $periode        = DB::table('periodes')->get();
         $jenis_pengdiri = DB::table('jenis_pengdiris')->get();
         // dd($periode);
@@ -60,6 +87,8 @@ class PengembanganController extends Controller
 
     public function update(Request $request, Pengembangan $pengembangan)
     {
+        $this->authorize('update', Pengembangan::class);
+
         $pengembangan = Pengembangan::findOrFail($pengembangan->id);
 
         $pengembangan->update([
@@ -76,6 +105,8 @@ class PengembanganController extends Controller
 
     public function destroy(Pengembangan $pengembangan)
     {
+        $this->authorize('delete', Pengembangan::class);
+
         $pengembangan->find($pengembangan->id)->all();
 
         $pengembangan->delete();
